Update codingstyle: class/filename, indentation
