+++ include/MeterMap.hpp
@@ -40,88 +40,88 @@
 #include <vector>
 
 #include <common.h>
-#include <options.h>
-#include <meter.h>
-#include <channel.h>
+#include <Options.hpp>
+#include <Meter.hpp>
+#include <Channel.hpp>
 
 /**
-   The MeterMap is intend to keep the list of all configured channel for a given meter.
- */
+	 The MeterMap is intend to keep the list of all configured channel for a given meter.
+*/
 class MeterMap {
-  public:
-  typedef vz::shared_ptr<MeterMap> Ptr;
-  typedef std::vector<Channel>::iterator iterator;
-  typedef std::vector<Channel>::const_iterator const_iterator;
+public:
+	typedef vz::shared_ptr<MeterMap> Ptr;
+	typedef std::vector<Channel>::iterator iterator;
+	typedef std::vector<Channel>::const_iterator const_iterator;
 
-  MeterMap(std::list<Option> options) : _meter(new Meter(options)){}
-  ~MeterMap() {};
-  Meter::Ptr meter() { return _meter; }
+	MeterMap(std::list<Option> options) : _meter(new Meter(options)){}
+	~MeterMap() {};
+	Meter::Ptr meter() { return _meter; }
 
-  /**
-     If the meter is enabled, start the meter and all its channels.
-   */
-  void start();
-    
-  /**
-     check if meter-thread is joinable
-   */
-  bool stopped();
-  
-  /**
-   * cancel all channels for this meter.
-   */
-  void cancel() {
-    pthread_cancel(_thread);
-    for(iterator it = _channels.begin(); it!=_channels.end(); it++) {
-      it->cancel();
-    }
-  }
-  
-  /** 
-   *  Accessor to the channel list
-   */
-  inline void push_back(Channel channel) { _channels.push_back(channel); }
-  inline iterator begin()  { return _channels.begin(); }
-  inline iterator end()    { return _channels.end(); }
-  inline size_t size()     { return _channels.size(); }
+/**
+	 If the meter is enabled, start the meter and all its channels.
+*/
+	void start();
 
-  private:
-  Meter::Ptr _meter;
+/**
+	 check if meter-thread is joinable
+*/
+	bool stopped();
+
+/**
+ * cancel all channels for this meter.
+ */
+	void cancel() {
+		pthread_cancel(_thread);
+		for(iterator it = _channels.begin(); it!=_channels.end(); it++) {
+			it->cancel();
+		}
+	}
+
+/** 
+ *  Accessor to the channel list
+ */
+	inline void push_back(Channel channel) { _channels.push_back(channel); }
+	inline iterator begin()  { return _channels.begin(); }
+	inline iterator end()    { return _channels.end(); }
+	inline size_t size()     { return _channels.size(); }
+
+private:
+	Meter::Ptr _meter;
 	std::vector<Channel> _channels;
 
 	pthread_t _thread;
 };
 
 /**
-   This container is intend to keep the list of all configured meters. 
- */
+	 This container is intend to keep the list of all configured meters. 
+*/
 class MapContainer {
-  public:
-  typedef vz::shared_ptr<MapContainer> Ptr;
-  typedef std::vector<MeterMap>::iterator iterator;
-  typedef std::vector<MeterMap>::const_iterator const_iterator;
+public:
+	typedef vz::shared_ptr<MapContainer> Ptr;
+	typedef std::vector<MeterMap>::iterator iterator;
+	typedef std::vector<MeterMap>::const_iterator const_iterator;
 
-  MapContainer() {};
-  ~MapContainer() {};
+	MapContainer() {};
+	~MapContainer() {};
 
-  void quit(int sig) {
-    print(log_info, "Closing connections to terminate", (char*)0);
+	void quit(int sig) {
+		print(log_info, "Closing connections to terminate", (char*)0);
 
-    for(iterator it = _mappings.begin(); it!=_mappings.end(); it++) {
-      it->cancel();
-    }
-  }
-  
-  /** 
-   *  Accessor to the MeterMap (meter and its channels) list
-   */
-  inline void push_back(const MeterMap &map) { _mappings.push_back(map); } 
-  inline iterator begin()  { return _mappings.begin(); }
-  inline iterator end()    { return _mappings.end(); }
-  inline const size_t size() const { return _mappings.size(); }
-  
-  private:
-  std::vector<MeterMap> _mappings;
-  
+		for(iterator it = _mappings.begin(); it!=_mappings.end(); it++) {
+			it->cancel();
+		}
+	}
+
+/** 
+ *  Accessor to the MeterMap (meter and its channels) list
+ */
+	inline void push_back(const MeterMap &map) { _mappings.push_back(map); } 
+	inline iterator begin()  { return _mappings.begin(); }
+	inline iterator end()    { return _mappings.end(); }
+	inline const size_t size() const { return _mappings.size(); }
+
+private:
+	std::vector<MeterMap> _mappings;
+
 };
 #endif /* _MeterMap_hpp_ */
