+++ src/api/MySmartGrid.cpp
@@ -37,7 +37,7 @@
 #include <openssl/evp.h>
 
 #include <VZException.hpp>
-#include "include/config.h"
+#include "Config_Options.hpp"
 #include <api/MySmartGrid.hpp>
 #include <api/CurlCallback.hpp>
