+++ src/protocols/MeterS0.cpp
@@ -30,12 +30,13 @@
 #include <sys/time.h>
 #include <errno.h>
 
-#include "protocols/s0.h"
-#include "options.h"
+#include "protocols/MeterS0.hpp"
+#include "Options.hpp"
 #include <VZException.hpp>
 
 MeterS0::MeterS0(std::list<Option> options)
 		: Protocol("s0", options)
+		, _counter(0)
 {
 	OptionList optlist;
 
@@ -49,11 +50,12 @@ MeterS0::MeterS0(std::list<Option> options)
 	try {
 		_resolution = optlist.lookup_int(options, "resolution");
 	} catch( vz::OptionNotFoundException &e ) {
-		_resolution = 1;
+		_resolution = 1000;
 	} catch( vz::VZException &e ) {
 		print(log_error, "Failed to parse resolution", "");
 		throw;
 	}
+	if(_resolution < 1) throw vz::VZException("Resolution must be greater than 0.");
 }
 
 MeterS0::~MeterS0() {
@@ -102,6 +104,8 @@ int MeterS0::close() {
 
 size_t MeterS0::read(std::vector<Reading> &rds, size_t n) {
 
+	struct timeval time1;
+	struct timeval time2;
 	char buf[8];
 
 	/* clear input buffer */
@@ -109,11 +113,29 @@ size_t MeterS0::read(std::vector<Reading> &rds, size_t n) {
 
 	/* blocking until one character/pulse is read */
 	if( ::read(_fd, buf, 8) < 1) return 0;
+	gettimeofday(&time1, NULL);
+	if( ::read(_fd, buf, 8) < 1) return 0;
+	gettimeofday(&time2, NULL);
+
+	double t1 = time1.tv_sec + time1.tv_usec / 1e6;
+	double t2 = time2.tv_sec + time2.tv_usec / 1e6;
+	double value = ( 3600000 ) / ( (t2-t1) * _resolution ) ;
 	
 	/* store current timestamp */
-	rds[0].time();
-	rds[0].value(1);
+	ReadingIdentifier *rid(new NilIdentifier());
+	rds[0].identifier(new StringIdentifier("Counter"));
+	rds[0].time(time2);
+	rds[0].value(++_counter);
 
+	rds[1].identifier(rid);
+	rds[1].time(time2);
+	rds[1].value(1);
+
+	rds[2].identifier(new StringIdentifier("Power"));
+	rds[2].time(time2);
+	rds[2].value(value);
+
+	print(log_debug, "Reading S0 - n=%d power=%f  counter=%f", name().c_str(), n, value, rds[2].value());
 	/* wait some ms for debouncing */
 	usleep(30000);
